+++ ckeditor5_plugins/callout/src/index.js
@@ -1,0 +1,5 @@
+import Callout from "./callout";
+
+export default {
+  Callout
+};