+++ ckeditor5_plugins/countup/src/index.js
@@ -1,0 +1,5 @@
+import Countup from "./countup";
+
+export default {
+  Countup
+};